+++ querybook/server/lib/ai_assistant/ai_assistant.py
@@ -1,47 +1,8 @@
 import threading
-import queue
-
-from langchain.callbacks.streaming_stdout import StreamingStdOutCallbackHandler
 
 
 from .all_ai_assistants import get_ai_assistant_class
-
-
-class ThreadedGenerator:
-    """Generator to facilitate streaming result from Langchain."""
-
-    def __init__(self):
-        self.queue = queue.Queue()
-
-    def __iter__(self):
-        return self
-
-    def __next__(self):
-        item = self.queue.get()
-        if item is StopIteration:
-            raise item
-        return item
-
-    def send(self, data):
-        self.queue.put(data)
-
-    def close(self):
-        self.queue.put(StopIteration)
-
-
-class ChainStreamHandler(StreamingStdOutCallbackHandler):
-    """Callback handlder to stream the result to a generator."""
-
-    def __init__(self, gen: ThreadedGenerator):
-        super().__init__()
-        self.gen = gen
-
-    def on_llm_new_token(self, token: str, **kwargs):
-        self.gen.send(token)
-
-    def on_llm_end(self, response, **kwargs):
-        self.gen.send(StopIteration)
-        self.gen.close()
+from .base_ai_assistant import ThreadedGenerator, ChainStreamHandler
 
 
 class AIAssistant:
