+++ querybook/server/lib/ai_assistant/assistants/openai_assistant.py
@@ -3,6 +3,11 @@ from lib.logger import get_logger
 
 from langchain.chat_models import ChatOpenAI
 from langchain.callbacks.manager import CallbackManager
+from langchain.prompts.chat import (
+    ChatPromptTemplate,
+    SystemMessagePromptTemplate,
+    HumanMessagePromptTemplate,
+)
 
 LOG = get_logger(__file__)
 
@@ -15,6 +20,22 @@ class OpenAIAssistant(BaseAIAssistant):
     @property
     def name(self) -> str:
         return "openai"
+
+    @property
+    def title_generation_prompt_template(self) -> str:
+        system_template = "You are a helpful assistant that can summerize SQL queries."
+        system_message_prompt = SystemMessagePromptTemplate.from_template(
+            system_template
+        )
+        human_template = (
+            "Generate a concise summary with no more than 8 words for the query below. "
+            "Only respond the title without any explanation or leading words.\n"
+            "```\n{query}\n```\nTitle:"
+        )
+        human_message_prompt = HumanMessagePromptTemplate.from_template(human_template)
+        return ChatPromptTemplate.from_messages(
+            [system_message_prompt, human_message_prompt]
+        )
 
     def generate_sql_query(self):
         pass
