--- requirements/ai/openai.txt
@@ -1,2 +1,0 @@
--r base.txt
-openai==0.27.5
